requirements.txt
python-gnupg
pyperclip
klembord